requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
beautifulsoup4==4.12.2
discord.py==2.3.2
# discord.py==2.1.0
Pillow==10.0.0
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.31.0